+++ ESLSCDLL/Board_ll.h
@@ -29,14 +29,14 @@ extern pthread_mutex_t mutex[MAXPCIECARDS];
 // platform specific implementation
 
 es_status_codes checkDriverHandle(uint32_t drvno);
-es_status_codes readRegister_32( uint32_t drvno, uint32_t* data, uint16_t address );
-es_status_codes readRegister_16( uint32_t drvno, uint16_t* data, uint16_t address );
-es_status_codes readRegister_8( uint32_t drvno, uint8_t* data, uint16_t address );
-es_status_codes writeRegister_32( uint32_t drvno, uint32_t data, uint16_t address );
-es_status_codes writeRegister_16( uint32_t drvno, uint16_t data, uint16_t address );
-es_status_codes writeRegister_8( uint32_t drvno, uint8_t data, uint16_t address );
-es_status_codes readConfig_32( uint32_t drvno, uint32_t* data, uint16_t address );
-es_status_codes writeConfig_32( uint32_t drvno, uint32_t data, uint16_t address );
+es_status_codes readRegister_32( uint32_t drvno, uint32_t* data, uint32_t address );
+es_status_codes readRegister_16( uint32_t drvno, uint16_t* data, uint32_t address );
+es_status_codes readRegister_8( uint32_t drvno, uint8_t* data, uint32_t address );
+es_status_codes writeRegister_32( uint32_t drvno, uint32_t data, uint32_t address );
+es_status_codes writeRegister_16( uint32_t drvno, uint16_t data, uint32_t address );
+es_status_codes writeRegister_8( uint32_t drvno, uint8_t data, uint32_t address );
+es_status_codes readConfig_32( uint32_t drvno, uint32_t* data, uint32_t address );
+es_status_codes writeConfig_32( uint32_t drvno, uint32_t data, uint32_t address );
 void FreeMemInfo( uint64_t *pmemory_all, uint64_t *pmemory_free );
 es_status_codes SetupDma( uint32_t drvno );
 es_status_codes enableInterrupt( uint32_t drvno );
